update: SkillSectionを更新
各スキルアイコンをコンポーネントに切り出しました。
SkillSectionのディレクトリ配置を変更しました。
SkillSectionのデザインを変更しました。
